backend/requirements.txt
aiofiles==23.2.1
chroma-hnswlib==0.7.6
chromadb==0.5.5
fastapi==0.111.0
langchain==0.2.10
langchain-community==0.2.10
langchain-core==0.2.23
langchain-openai==0.1.17
openai==1.37.0
pandas==2.2.1
python-docx==1.1.2
starlette==0.37.2
uvicorn==0.29.0

